# swagger.yaml
openapi: 3.0.0
info:
  title: kubuka kwesu tobecisi cabutonga
  description: Kubuka is an organisation
  version: 1.0.0
servers:
  - url: https://server.kubukaspace.com
    description: Online development server
paths:
...
  /cart/{id}/create:
    post:
      summary: Book for your venue and meals
      description: The booking office is not responsible for loss of materials during events
      operationId: createCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                  example: "Getme"
                company_address:
                  type: string
                  example: "Male 2355 mkhb3"
                agent_name:
                  type: string
                  example: "Mape Gweru"
                agent_phone:
                  type: string
                  example: "04411552320"
                numberOfDays:
                  type: number
                  example: 5
                agent_email:
                  type: string
                  example: "user01@kubuka.com"
                numberOfGuests:
                  type: number
                  example: 40
                from_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                to_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                venue_type:
                  type: string
                  example: "Boord room"
                price:
                  type: number
                  example: 200
              required:
                - company_name
                - company_address
                - agent_name
                - agent_email
                - agent_phone
                - numberOfDays
                - numberOfGuests
                - from_date
                - to_date
                - venue_type
                - price
      responses:
        "200":
          description: The Cart was made to the organisation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cart was successfully made"
  /cart/{uid}:
    parameters:
      - in: path
        name: id
        description: Get all items in cart
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single cart by id
      description: Admin users can access any cart, regular users are restricted to their own cart.
      operationId: getAllCartsById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  company_name:
                    type: string
                    example: "Getme"
                  company_address:
                    type: string
                    example: "45 manho gerus"
                  agent_name:
                    type: string
                    example: "minuhed Mig"
                  agent_phone:
                    type: string
                    example: "0412251745"
                  numberOfDays:
                    type: number
                    example: 5
                  agent_email:
                    type: string
                    example: "user01@kubuka.com"
                  numberOfGuests:
                    type: number
                    example: 4
                  venue_type:
                    type: string
                    example: "Igzin for"
                  price:
                    type: number
                    example: 2000
                  cartOwnerId:
                    type: string
                    example: "51eb12e197e06a76cdcdefc121"
                  from_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  to_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  cancelled:
                    type: boolean
                    example: true
                  honoured:
                    type: boolean
                    example: true
                  returned:
                    type: boolean
                    example: true
                  confirmed:
                    type: boolean
                    example: true
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /cart/{iid}/{uid}/update:
    parameters:
      - in: path
        name: id
        description: Cart Update schema
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    put:
      summary: Update a cart
      description: Owner of the cart can update, admins can also update certain parts only
      operationId: updateCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                  example: "Getme"
                company_address:
                  type: string
                  example: "Male 2355 mkhb3"
                agent_name:
                  type: string
                  example: "Mape Gweru"
                agent_phone:
                  type: string
                  example: "04411552320"
                numberOfDays:
                  type: number
                  example: 5
                agent_email:
                  type: string
                  example: "user01@kubuka.com"
                numberOfGuests:
                  type: number
                  example: 40
                from_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                to_date:
                  type: string
                  example: "2020-05-05T09:12:57.848Z"
                venue_type:
                  type: string
                  example: "Boord room"
                price:
                  type: number
                  example: 200
                cancelled:
                  type: boolean
                  example: true
                honoured:
                  type: boolean
                  example: true
                returned:
                  type: boolean
                  example: true
                confirmed:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Account updated successfully. The details of the updated account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  company_name:
                    type: string
                    example: "Getme"
                  company_address:
                    type: string
                    example: "45 manho gerus"
                  agent_name:
                    type: string
                    example: "minuhed Mig"
                  agent_phone:
                    type: string
                    example: "0412251745"
                  numberOfDays:
                    type: number
                    example: 5
                  agent_email:
                    type: string
                    example: "user01@kubuka.com"
                  numberOfGuests:
                    type: number
                    example: 4
                  venue_type:
                    type: string
                    example: "Igzin for"
                  price:
                    type: number
                    example: 2000
                  cartOwnerId:
                    type: string
                    example: "51eb12e197e06a76cdcdefc121"
                  from_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  to_date:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  cancelled:
                    type: boolean
                    example: true
                  honoured:
                    type: boolean
                    example: true
                  returned:
                    type: boolean
                    example: true
                  confirmed:
                    type: boolean
                    example: true
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /cart/{iid}/{uid}:
    parameters:
      - in: path
        name: id
        description: Cart Item delete
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    delete:
      summary: Delete a cart
      description: Admin users can delete any cart, regular users are restricted to their own cart.
      operationId: deleteCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cart deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /comments/{id}/comment:
    post:
      summary: Write a new comment
      description: The comments to poems
      operationId: newComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                commenter_name:
                  type: string
                  example: "Author Writer"
                comment_body:
                  type: string
                  example: "Get me that"
              required:
                - commenter_name
                - comment_body
      responses:
        "200":
          description: Comment has been saved to database
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment has been saved to database"
  /comments/{bid}:
    parameters:
      - in: path
        name: id
        description: Comment blog id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a list of all comments
      description: All users, including guests, can access this information.
      operationId: getAllComments
      responses:
        "200":
          description: An array of all comments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    commenter_name:
                      type: string
                      example: "Author"
                    comment_body:
                      type: string
                      example: "The poem"
                    comment_to_id:
                      type: string
                      example: "5ced334ab4554bbafdde445da4ed2"
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /comments/{bid}/{cid}:
    parameters:
      - in: path
        name: id
        description: Comment id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single comment by id
      description: Admin users can access any comment, regular users are restricted to their own comment.
      operationId: getAccountById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  comment_body:
                    type: string
                    example: "Mr me and my"
                  commenter_name:
                    type: string
                    example: "Author"
                  comment_to_id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /comments/{id}:
    parameters:
      - in: path
        name: id
        description: Comment id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    put:
      summary: Update a comment
      description: Admin users can update any comment
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                commenter_name:
                  type: string
                  example: "Author"
                comment_body:
                  type: string
                  example: "The comment"
      responses:
        "200":
          description: Comment updated successfully. The details of the updated comment are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  commenter_name:
                    type: string
                    example: "Author"
                  comment_body:
                    type: string
                    example: "The poem"
                  comment_to_id:
                    type: string
                    example: "5ced334ab4554bbafdde445da4ed2"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete an Comment
      description: Admin users can delete items
      operationId: deleteComment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"